Programmed the HighScore class. It uses a hashmap to store an ArrayList which contains the name of the player and the time. The hashmap uses as the key the difficulty name. The hashmap is then stored on the hard drive in XML. When the HighScore Window is called, it tries to find the HighScores.xml file, and if it doesn't find it, or if there is an error in it, a new file is initialized. Also implemented the methods to save the HighScores to file, to set the highScore, to delete the HighScores and to compare a HighScore.
With the HighScore class programmed, I modified HighScoresWindowController to use it. When the window is initialized, it loads the HighScores. It is
also possible to delete the highScores and to set a HighScore. There is also a temp method "setTestHighScore" used to test if everything is working
correctly.

Finally, I removed the TableView from the HighScoresWindow.fxml file because to use the TableView I had to modify way to many things to make it work,
so I decided to go for a simple (albeit a bit more ugly) solution: Using a 3x3 gridpane with labels. And when you think about it, using a TableView
does not make that much sens, because the number of colums and rows is a fixed number (3)... So we didn't need the flexibility of a TableView (if was
too OP for our purpose).

Yippy!